# new: dev: pos-620 introduce cache for contracts ABIs