@marcello33 2022-07-18
> @marcello33 could we fix comments, ask for the second CR and merge it?

Just pushed some changes to address the comments ;) waiting for CI to do its job